==> bin/18_random_forest.py <==
import pandas as pd
from sklearn.ensemble import RandomForestClassifier


def rf(df,):
    df = df
    X, y =
    rf = RandomForestClassifier()
    rf.fit()


def main():
    pass


if __name__ == '__main__':
    main()
